Struct lib_gst_meet::JitsiConference

pub struct JitsiConference { /* private fields */ }

Implementations

impl JitsiConference

pub async fn join( xmpp_connection: Connection, glib_main_context: MainContext, config: JitsiConferenceConfig ) -> Result<Self>

pub async fn leave(self) -> Result<()>

pub async fn set_muted(&self, media_type: MediaType, muted: bool) -> Result<()>

pub async fn pipeline(&self) -> Result<Pipeline>

pub async fn add_bin(&self, bin: &Bin) -> Result<()>

pub async fn set_pipeline_state(&self, state: State) -> Result<()>

pub async fn remote_participant_audio_sink_element(&self) -> Option<Element>

pub async fn set_remote_participant_audio_sink_element( &self, sink: Option<Element> )

pub async fn remote_participant_video_sink_element(&self) -> Option<Element>

pub async fn set_remote_participant_video_sink_element( &self, sink: Option<Element> )

pub async fn audio_sink_element(&self) -> Result<Element>

pub async fn video_sink_element(&self) -> Result<Element>

pub async fn set_send_resolution(&self, height: i32)

Set the max resolution that we are currently sending.

Setting this is required for browser clients in the same conference to display the stats that we broadcast.

Note that lib-gst-meet does not encode video (that is the responsibility of your GStreamer pipeline), so this is purely informational.

pub async fn send_colibri_message(&self, message: ColibriMessage) -> Result<()>

pub async fn send_json_message<T: Serialize>(&self, payload: &T) -> Result<()>

pub async fn on_participant( &self, f: impl Fn(JitsiConference, Participant) -> Pin<Box<dyn Future<Output = Result<()>> + Send>> + Send + Sync + 'static )

pub async fn on_participant_left( &self, f: impl Fn(JitsiConference, Participant) -> Pin<Box<dyn Future<Output = Result<()>> + Send>> + Send + Sync + 'static )

pub async fn on_colibri_message( &self, f: impl Fn(JitsiConference, ColibriMessage) -> Pin<Box<dyn Future<Output = Result<()>> + Send>> + Send + Sync + 'static )